Move tasks into packaged
