+++ src/ftl/tasks/__init__.py
@@ -1,5 +1,7 @@
 # tasks API
 from .base import ParameterizedTask, Task, parameterize
+from .encode_folder import EncodeFolder
 from .encode_gif import EncodeGif
 from .encode_mov import EncodeMov
 from .encode_mp4 import EncodeMp4
+from .types import Fps, Size
