+++ src/ftl/tasks/encode_mov.py
@@ -1,18 +1,93 @@
+import subprocess
 from dataclasses import dataclass, field
 from typing import Literal
 
+from ftl.files import FileSequence, PathType, as_file
 from ftl.tasks.base import Task
-from ftl.tasks.parameter_types import Fps, Size
+from ftl.tasks.types import Fps, Size
+from ftl.tools import get_ffmpeg
 
 CodecProres = Literal["Prores422", "Prores4444"]
 
 
 @dataclass
 class EncodeMov(Task):
+    src: PathType
+    dst: PathType
     input_colorspace: str = field(default="srgb", kw_only=True)
     vcodec: CodecProres = field(default="Prores4444", kw_only=True)
     fps: Fps = field(default=-1, kw_only=True)
     max_size: Size = field(default=-1, kw_only=True)
 
-    def run(self):
-        print(self)
+    def command(self) -> list[str]:
+        # fmt: off
+        cmd = [
+            get_ffmpeg(),
+            "-r", str(self.fps),
+            "-i", str(self.src),
+            "-c:v", "prores_ks",
+            "-profile:v", "4444",
+            "-qscale:v", "11",
+            "-vendor", "apl0",
+            "-pix_fmt", "yuva444p10le",
+            "-alpha_bits", "16",
+            "-color_range", "tv",
+            "-colorspace", "bt709",
+            "-color_primaries", "bt709",
+            "-color_trc", "iec61966-2-1",
+            "-y",
+            str(self.dst),
+        ]
+        # fmt: on
+
+        # Filter Graph
+        # Colorspace
+        filters = []
+        if self.src.suffix == ".exr":
+            filters.append(
+                "colorspace=space=bt709:primaries=bt709:trc=srgb:range=tv:ispace=bt709:iprimaries=bt709:itrc=linear:irange=tv"
+            )
+        else:
+            filters.append("scale=in_color_matrix=bt709:out_color_matrix=bt709")
+
+        # Scale
+        if self.max_size > 0:
+            filters.append(
+                f"scale='{self.max_size}:{self.max_size}:force_original_aspect_ratio=decrease:force_divisible_by=2'"
+            )
+
+        # Apply Filter Graph
+        if filters:
+            cmd[7:7] = [
+                "-vf",
+                ",".join(filters),
+            ]
+
+        # Ensure correct start_number for File Sequences
+        f = as_file(self.src)
+        if isinstance(f, FileSequence):
+            cmd[1:1] = ["-start_number", str(f.frame_start)]
+
+        return cmd
+
+    def run(self) -> PathType:
+        # Ensure destination directory exists...
+        self.dst.parent.mkdir(parents=True, exist_ok=True)
+
+        cmd = self.command()
+
+        self.log(f"  MOV: {self.src.name} -> {self.dst.name}")
+        self.log(f"  CMD: {' '.join(cmd)}\n")
+
+        try:
+            subprocess.run(
+                cmd,
+                check=True,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+                creationflags=subprocess.CREATE_NO_WINDOW,
+            )
+        except subprocess.CalledProcessError as e:
+            raise RuntimeError(f"FFmpeg failed with error code {e.returncode}") from e
+
+        return self.dst
