+++ src/ftl/tasks/types.py
@@ -1,4 +1,4 @@
 from typing import Literal
 
 Size = Literal[-1, 256, 512, 768, 1024, 1280, 1920, 2048, 2160, 3840, 6144]
-Fps = Literal[-1, 24, 25, 30, 48, 50, 60]
+Fps = Literal[-1, 8, 12, 15, 24, 25, 30, 48, 50, 60]
